requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
pymupdf==1.26.5
faiss-cpu==1.13.0
python-multipart==0.0.20
python-dotenv==1.2.1
requests==2.32.5
openai==2.30.0
google-genai==1.47.0
huggingface-hub==1.8.0
jinja2==3.1.6
streamlit==1.50.0
scipy==1.17.1
scikit-learn==1.8.0
transformers==4.57.6
tokenizers==0.22.2
safetensors==0.7.0
torch==2.11.0
sentence-transformers==5.3.0
pdfplumber==0.11.9
chromadb==1.5.5
pinecone-client==6.0.0
